+++ routers/googleCal-routes.js
@@ -2,7 +2,6 @@ const axios = require("axios");
 const request = require("request");
 const User = require("../models/user-model");
 const moment = require("moment");
-const route = require("./routers");
 
 require("dotenv").config();
 
@@ -42,7 +41,7 @@ module.exports = {
     request.post(options, (err, httpResonse, body) => {
       console.log("body--gCal------------\n>", body);
       if (body.error) {
-        route.refreshAccessToken(event, user);
+        refreshAccess(event, user);
       }
 
       // console.log("err--gCal--------------------\n>", err);
@@ -51,23 +50,23 @@ module.exports = {
   }
 };
 
-// const refreshAccessToken = async (event, user) => {
-//   let url = `https://www.googleapis.com/oauth2/v4/token?refresh_token=${
-//     user.google_refresh_token
-//     }&client_id=${process.env.GOOGLE_CLIENT_ID}&client_secret=${
-//     process.env.GOOGLE_CLIENT_SECRET
-//     }&grant_type=refresh_token`;
-//   await request.post(url, async (err, httpResponse, body) => {
-//     let data = JSON.parse(body);
+const refreshAccess = async (event, user) => {
+  let url = `https://www.googleapis.com/oauth2/v4/token?refresh_token=${
+    user.google_refresh_token
+  }&client_id=${process.env.GOOGLE_CLIENT_ID}&client_secret=${
+    process.env.GOOGLE_CLIENT_SECRET
+  }&grant_type=refresh_token`;
+  await request.post(url, async (err, httpResponse, body) => {
+    let data = JSON.parse(body);
 
-//     await User.updateOne(
-//       { _id: user._id },
-//       { google_access_token: data.access_token }
-//     );
-//     // if (updatedUser)
-//     //       googleRoutes.add_to_google(event);
-//   });
-// }
+    await User.updateOne(
+      { _id: user._id },
+      { google_access_token: data.access_token }
+    );
+    // if (updatedUser)
+    //       googleRoutes.add_to_google(event);
+  });
+};
 // POST https://www.googleapis.com/calendar/v3/calendars/[CALENDARID]/events?key=[YOUR_API_KEY] HTTP/1.1
 
 // Authorization: Bearer[YOUR_ACCESS_TOKEN]
