+++ routers/routers.js
@@ -217,25 +217,25 @@ module.exports = {
       google_refresh_token: refresh
     });
     await findUser.save();
-  },
-
-  refreshAccessToken: async (event, user) => {
-    let url = `https://www.googleapis.com/oauth2/v4/token?refresh_token=${
-      user.google_refresh_token
-    }&client_id=${process.env.GOOGLE_CLIENT_ID}&client_secret=${
-      process.env.GOOGLE_CLIENT_SECRET
-    }&grant_type=refresh_token`;
-    await request.post(url, async (err, httpResponse, body) => {
-      let data = JSON.parse(body);
-
-      const userToken = await User.updateOne(
-        { _id: user._id },
-        { google_access_token: data.access_token }
-      );
-
-      if (userToken.n === 1) {
-        google.add_to_google(event);
-      }
-    });
   }
+
+  // refreshAccessToken: async (event, user) => {
+  //   let url = `https://www.googleapis.com/oauth2/v4/token?refresh_token=${
+  //     user.google_refresh_token
+  //   }&client_id=${process.env.GOOGLE_CLIENT_ID}&client_secret=${
+  //     process.env.GOOGLE_CLIENT_SECRET
+  //   }&grant_type=refresh_token`;
+  //   await request.post(url, async (err, httpResponse, body) => {
+  //     let data = JSON.parse(body);
+
+  //     const userToken = await User.updateOne(
+  //       { _id: user._id },
+  //       { google_access_token: data.access_token }
+  //     );
+
+  //     if (userToken.n === 1) {
+  //       google.add_to_google(event);
+  //     }
+  //   });
+  // }
 };
